🔒️ (maxGas NDS bounds) Bounds set in initializers

+++ contracts/aave/positions-manager-parts/PositionsManagerForAaveEventsErrors.sol
@@ -106,7 +106,7 @@ abstract contract PositionsManagerForAaveEventsErrors is PositionsManagerForAave
 
     /// @dev Emitted when a new value for `NDS` is set.
     /// @param _newValue The new value of `NDS`.
-    event NDSSet(uint8 _newValue);
+    event NDSSet(uint32 _newValue);
 
     /// @dev Emitted when a new `maxGas` is set.
     /// @param _maxGas The new `maxGas`.
